backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cross-web==0.4.0
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.1.0
graphql-core==3.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
kombu==5.6.2
openai==2.15.0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
strawberry-graphql==0.288.1
strawberry-graphql-django==0.72.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
